LayoutTests/fast/forms/datetimelocal/datetimelocal-editable-components/up-down-arrow-starts-at-current-year-expected.txt

Tests that pressing the up or down arrow on the year field of a date input starts the value at the current year.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS new Date(input1.value).getFullYear() === currentYear + 1 is true
PASS new Date(input2.value).getFullYear() === currentYear - 1 is true
PASS successfullyParsed is true

TEST COMPLETE
